partial-register
